Add files via upload
Problem – Output your favorite three games

Launch the Unreal Editor, select Games as the New Project Category and click the Next button. Select the Blank Template and click the Next button. In the Project Settings, select a C++ Project on the upper left and select No Starter Content on the middle right. Select a Folder and enter Exercise2 as the Name for the project at the bottom. Click the Create Project button near the bottom right.

If the project creation process brought you to your IDE (Visual Studio or Xcode), return to the Unreal Editor. Right click in the Content Browser panel on the bottom and select New Folder near the top. Rename the new folder Exercise2 (We’re generally following good Unreal Engine development folder structure; see https://github.com/Allar/ue4-style-guide for additional details). Double click the new folder to open it, then add a folder named Maps using the same process.

Click the Save Current button in the toolbar near the top middle of the editor (or use Ctrl+S) to save the current map. Navigate to the Maps folder in the popup, double click it, change the name at the bottom of the popup to Map0, then click the Save button near the bottom right of the popup.

Now we'll make it so this map opens up automatically when you (or someone else) opens up this project. Select Edit > Project Settings from the top menu bar. Click Maps & Modes in the pane on the left. In the Default Maps section on the right, Set both the Editor Startup Map and the Game Default Map to Map0 by clicking the dropdown to the right of the map icon and selecting Map0. Click the x in the upper right corner of the Project Settings window to close it.

On the top menu, select File > New C++ Class... In Choose Parent Class, select Actor and click the Next button. Change the Name to PrintFavoriteGames. Leave the Path as the default path and click the green Create Class button near the bottom right.

Open the script in your IDE. It may automatically open for you, but if it doesn’t, just move over to your IDE and tell it to reload the project if necessary. Make sure you’re editing the .cpp file, not the .h file. In the constructor (the function with the comment that includes Sets default value above it), change true in the line that says PrimaryActorTick.bCanEverTick = true; to false. This makes it so the Tick function won’t be called on every frame;  we only need to print our message once when the game starts, not on every frame of the game.

Add code to the body of the BeginPlay  function (the function with the comment that includes Called when the game starts or when spawned above it) to output the names of your three favorite games. If one of your favorite games is Assetto Corsa, here’s the code you would add below the line that says Super::BeginPlay();:

1
UE_LOG(LogTemp, Warning, TEXT("Assetto Corsa"));
Strictly speaking, UE_LOG is a macro, but we can use it just like a function. UE_LOG outputs information to the Output Log window, which is just what we want. In the line of code above, LogTemp is the log category. Although we can define our own log categories, using the existing LogTemp category will work fine. Warning is called the verbosity level; it determines how much information is included in the log message and what color it is. Finally, TEXT("Assetto Corsa") provides the text we want to be displayed in the log message.

Navigate to your new script in the Content Browser panel by clicking the folder icon to the left of Content, expanding the C++ Classes folder if necessary (click the arrow to the left of it), and selecting the Exercise2 folder. Drag the PrintFavoriteGames script from the Content Brower panel into the Viewport and drop it there. Ctrl+S to save the map.

Now we need to make the output log window visible so we can see our log messages. Go to the top menu bar and select Window > Developer Tools > Output Log.

Press Alt+P to play the game; you should see your three games output in yellow text in the Output Log window (you may have to scroll up to find them). Press the Escape key to stop playing the game.